feat: implement MonitorValidator for monitor configuration validation
